+++ src/components/views/liveview/LiveviewBars.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { LiveDataProgress } from './LiveDataProgress';
-import { Grid, Col, Row } from './ui/Grid';
+import { Grid, Col, Row } from '../../ui/Grid';
 
 export const LiveviewBars = (props) => {
     return (
