+++ src/components/views/scheduleview/SchedulingList.js
@@ -1,8 +1,8 @@
 import React, { useState }  from 'react'
 import styled from 'styled-components';
 import { Icon } from '@iconify/react-with-api';
-import { Table, Head, SortableBody } from './ui/Tables';
-import { DeleteButton } from './ui/Buttons';
+import { Table, Head, SortableBody } from '../../ui/Tables';
+import { DeleteButton } from '../../ui/Buttons';
 
 export const SchedulingList = (props) => {
 	const [selected, setSelected] = useState('');
